% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eff_dens.R
\name{calc_eff_dens}
\alias{calc_eff_dens}
\title{Calculate effective densities}
\usage{
calc_eff_dens(gen_costs, jobs, year)
}
\arguments{
\item{gen_costs}{a dataframe of generalised costs in the following format:
`origin_zone`, `destination_zone`, `generalised_cost`}

\item{jobs}{a dataframe of jobs in each sector for each modelled zone. If the
analysis is being undertaken using the TII National Transport Model, the
\code{\link{ntpm_jobs}} data file provided with the `pagglomR` package can be
used. For other models, this file needs to be prepared for the transport
zone system in question.}

\item{year}{the modelled year. It is important that this is correctly
specified for the purposes of the productivity calculations and subsequent
discounting required.}
}
\value{
A list containing the modelled year and an effective density matrix
that are required for the productivity calculations.
}
\description{
This function reads in a list of generalised costs by origin-destination pair
and also a dataframe of jobs by sector. It calculates the effective densities
for each modelled zone and for each employment sector.
}
\keyword{agglomeration,}
\keyword{density}
\keyword{effective}
