% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod_impacts.R
\name{calc_prod_impacts}
\alias{calc_prod_impacts}
\title{Calculate effective densities}
\usage{
calc_prod_impacts(eff_dens_dm, eff_dens_ds, jobs)
}
\arguments{
\item{eff_dens_dm}{Do Minimum effective density object create by
\code{\link{calc_eff_dens}}}

\item{eff_dens_ds}{Do Something effective density object create by
\code{\link{calc_eff_dens}}}

\item{jobs}{dataframe of jobs}
}
\value{
A list containing the modelled year and a productivity matrix
}
\description{
This function reads in Do Minimum and Do Something effective density matrices,
and jobs data and applies the PAG agglomeration parameters to calculate
producvity impacts for the inputted model year
}
\keyword{agglomeration,}
\keyword{productivity}
