% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discounted_prod.R
\name{discounted_prod}
\alias{discounted_prod}
\title{Produce table of modelled, nominal and discounted productivity benefits}
\usage{
discounted_prod(...)
}
\arguments{
\item{...}{objects containing the productivity benefits for each modelled
year output using the \code{calc_prod_impacts} function.}
}
\value{
A table of benefits over the 60 year appraisal period
}
\description{
This function generates a table of the modelled, nominal and discounted
productivity benefits over the 60 year appraisal period. It reads in the
producitivty benefits calculated for each modelled year and interpolates
values for years in between the modelled years. It also calculates nominal
productivity benefits per annum by applying GVA growth based on the rates in
Table X of PAG Unit 6.11. Discounted benefits are then calculated by applying
the test discount rate to the nominal benefits.
}
\keyword{agglomeration,}
\keyword{benefits,}
\keyword{discounting}
