% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dm_2039}
\alias{dm_2039}
\title{Do Minimum 2039}
\format{
A tibble with 3 variables: \code{o_zone}: Origin Zone,
\code{d_zone}: Destination Zone, \code{gen_cost}: generalised cost for each
OD pair
}
\usage{
dm_2039
}
\description{
Sample generalised cost outputs from a transport model for a 2039 forecast
year and a Do Minimum scenario.
}
\keyword{datasets}
