% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ds_2054}
\alias{ds_2054}
\title{Do Something 2054}
\format{
A tibble with 3 variables: \code{o_zone}: Origin Zone,
\code{d_zone}: Destination Zone, \code{gen_cost}: generalised cost for each
OD pair
}
\usage{
ds_2054
}
\description{
Sample generalised cost outputs from a transport model for a 2054 forecast
year and a Do Something scenario.
}
\keyword{datasets}
