% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gva_factors}
\alias{gva_factors}
\title{GVA Factors}
\format{A data frame with fout variables: \code{year}: year,
 \code{discount}: discount rate, \code{gva_growth}: GVA growth rates from
 PAG, \code{gva_compound}: GVA growth compounded}
\usage{
gva_factors
}
\description{
Factors to apply growth to future year GVA values and to discount to present
value
}
\keyword{datasets}
