% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ntpm_jobs}
\alias{ntpm_jobs}
\title{NTpM Jobs}
\format{
A dataframe with 6 variables: \code{Manufacturing}: jobs
in manufacturing sector, \code{Construction}: jobs in construction sector,
\code{Wholesale_Retail}: jobs in wholesale & retail sector,
\code{Inf_Comm_Tech}: jobs in information, communication & technology
 sector, #' \code{Fin_Bus_Services}: jobs in financial & business services
 sector.
}
\usage{
ntpm_jobs
}
\description{
National jobs data from the TII National Transport Model (NTpM) for use with
the \code{pagglomR} package. Note that this is only compatible with the NTpM
zone system and not with any other transport model.
}
\keyword{datasets}
