% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{parameters}
\alias{parameters}
\title{Agglomeration Parameters}
\format{A data frame with 2 variables: \code{sector}: Sector,
\code{gva_person}: GVA per person, \code{decay_parameter}: distance decay
parameter, \code{elasticity}: elasticity of effictive density to productivity}
\usage{
parameters
}
\description{
Sectoral agglomeration parameters
}
\keyword{datasets}
