% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prod_summary.R
\name{prod_summary}
\alias{prod_summary}
\title{Produce summary table of productivity benefits}
\usage{
prod_summary(prod_table)
}
\arguments{
\item{prod_table}{objects containing the productivity benefits for each
modelled year output using the \code{calc_prod_impacts} function.}
}
\value{
A productivity benefits summary table
}
\description{
This function generates a summary table of the productivity benefits in
present value terms using an input object generated from the
\code{\link{discounted_prod}} function. These benefits are expressed in
present value terms for the 30 appraisal period and the residual value
period. They can be added to the Present Value of Benefits (PVB) from the
transport user benefits element of the Cost Benefit Analysis as part of the
sensitivity testing process.
}
\keyword{agglomeration,}
\keyword{benefits,}
\keyword{discounting}
