% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_jobs.R
\name{read_jobs}
\alias{read_jobs}
\title{Read in jobs data}
\usage{
read_jobs(path)
}
\arguments{
\item{path}{path to where the jobs data is saved}
}
\value{
A dataframe of jobs
}
\description{
This function reads in tab delimited formatted jobs data
}
\keyword{agglomeration,}
\keyword{jobs}
