% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sample_jobs}
\alias{sample_jobs}
\title{Sample Jobs}
\format{
A dataframe with 6 variables: \code{Manufacturing}: jobs
in manufacturing sector, \code{Construction}: jobs in construction sector,
\code{Wholesale_Retail}: jobs in wholesale & retail sector,
\code{Inf_Comm_Tech}: jobs in information, communication & technology
 sector, #' \code{Fin_Bus_Services}: jobs in financial & business services
 sector.
}
\usage{
sample_jobs
}
\description{
Sample jobs data for demo of \code{pagglomR} package.
}
\keyword{datasets}
