% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_eff_dens.R
\name{calc_eff_dens}
\alias{calc_eff_dens}
\title{Calculate effective densities}
\usage{
calc_eff_dens(gen_costs, jobs, year)
}
\arguments{
\item{gen_costs}{dataframe of generalised costs}

\item{jobs}{dataframe of jobs}

\item{year}{modelled year}
}
\value{
A list containing the modelled year and an effective density matrix
}
\description{
This function reads in a matrix of generalised costs and calculates the
effective densities for each origin destination pair
}
\keyword{agglomeration,}
\keyword{density}
\keyword{effective}
