% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_prod_impacts.R
\name{calc_prod_impacts}
\alias{calc_prod_impacts}
\title{Calculate effective densities}
\usage{
calc_prod_impacts(eff_dens_dm, eff_dens_ds, jobs)
}
\arguments{
\item{eff_dens_dm}{Do Minimum effective density object create by
\code{\link{calc_eff_dens}}}

\item{eff_dens_ds}{Do Something effective density object create by
\code{\link{calc_eff_dens}}}

\item{jobs}{a dataframe of jobs in each sector for each modelled zone. If the
analysis is being undertaken using the TII National Transport Model, the
\code{\link{jobs}} data file provided with the `pagglomR` package can be
used. For other models, this file needs to be prepared for the transport
zone system in question.}
}
\value{
A list containing the modelled year and a productivity matrix
}
\description{
This function reads in Do Minimum and Do Something effective density matrices,
and jobs data and applies the TII PAG Irish agglomeration parameters to
calculate productivity impacts for the relevant modelled year.
}
\keyword{agglomeration,}
\keyword{productivity}
