% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export_results.R
\name{export_results}
\alias{export_results}
\title{Output productivity impacts to an MS Word report or a csv file}
\usage{
export_results(
  benefits_summary,
  benefits_discounted,
  output_format = c("report", "csv"),
  output_file,
  scheme_name,
  report_date
)
}
\arguments{
\item{benefits_summary}{the \code{R} object output by the
\code{\link{prod_summary}} function}

\item{benefits_discounted}{the \code{R} object output by the
\code{\link{discounted_prod}} function}

\item{output_format}{specify the output format as "report" for an MS Word
report or "csv" for a csv table of the discounted benefits by year.}

\item{scheme_name}{input the scheme name to be included in the MS Word report}

\item{report_date}{input the date to be included in the MS Word report}
}
\value{
A table of benefits over the 60 year appraisal period
}
\description{
This function takes the outputs of the \code{\link{discounted_prod}} and the
\code{\link{prod_summary}} functions to produce an MS Word report or csv
summary of the outputs.
}
\keyword{agglomeration,}
\keyword{benefits,}
\keyword{output,}
\keyword{report,}
\keyword{results}
